tquic: fix 106 issues from deep security audit round 2
Seven parallel audit agents analyzed the codebase for:
  - Integer overflow/underflow (15 findings)
  - Use-after-free patterns (14 findings)
  - State machine correctness (16 findings)
  - Denial-of-service attack surface (16 findings)
  - Error path resource leaks (12 findings)
  - RFC 9000/9001/9002 compliance (16 findings)
  - Data races and atomicity violations (17 findings)

Critical fixes (17):
  - check_add_overflow() for coalesced packet token/length parsing
  - Path list and stream RB-tree lookups now hold proper locks
  - MP_ACK keeps paths_lock during ack_state access
  - Frame filtering in CLOSING/DRAINING states (RFC 9000 §10.2)
  - Stream state machine validates all transitions from OPEN
  - Stream reset goes through state validator
  - MAX_STREAMS default reduced from 2^60 to 256
  - Incoming streams validated against MAX_STREAMS limits
  - Header protection stubs replaced with real crypto integration
  - Unknown frame types trigger CONNECTION_CLOSE with FRAME_ENCODING_ERROR
  - conn->state writes protected by conn->lock in handshake callbacks
  - path->state uses consistent WRITE_ONCE/READ_ONCE across subsystems
  - Flow control check-and-update atomic under conn->lock (TOCTOU fix)
  - SKB leak fixed in handshake crypto_buffer on kzalloc failure

High fixes (29):
  - Anti-amplification bytes_sent/bytes_received now atomic64_t
  - Stateless reset token: jhash replaced with HMAC-SHA256
  - Key update uses staging approach to eliminate lock gap
  - Retry AEAD cached (not allocated per-packet)
  - Retry token AEAD pool (8 slots) eliminates mutex contention
  - Timer work takes conn refcount before using connection
  - Stream socket ops take conn refcount
  - Failover destroy always calls cancel_delayed_work_sync
  - Failover on_ack retx_queue TOCTOU eliminated
  - All bonding/failover stats counters now atomic64_t
  - NETDEV_DOWN handler holds paths_lock during iteration
  - Connection work cancellation: sync variants outside work context
  - CID list iteration under conn->lock
  - PATH_CHALLENGE rate limited to 4/RTT
  - cwnd * USEC_PER_SEC overflow capped in all CC algorithms
  - ACK delay shift overflow capped per RFC 9000 §18.2
  - ACK range underflow guards
  - FEC allocation bounds checked with kmalloc_array
  - Drain timeout computed from actual PTO, not static default
  - Encryption level enforcement per RFC 9000 §12.4 Table 3
  - HANDSHAKE_DONE rejected by servers per RFC 9000 §19.20

Medium fixes (36):
  - Per-CPU loss detection replaced with per-connection
  - PTO calculation space-aware (max_ack_delay only for app data)
  - Flow control subtraction-based overflow protection
  - HKDF buffers zeroized on all error paths
  - krealloc capacity overflow protected in cert_verify
  - Weight update flag uses test_and_set_bit
  - Bonding state re-evaluated after lock drop/reacquire
  - READ_ONCE for conn->active_path in proc/diag/proto
  - Netns exit uses refcount-driven connection destroy
  - Retry version defaults return -EPROTONOSUPPORT
  - Per-IP rate limit table capped at 100K entries
  - ACK frequency parameters bounds-checked
  - Netlink path_set holds conn->lock
  - IMMEDIATE_ACK deduped (one per packet)
  - CRYPTO frame pre-handshake memory limit enforced
  - Migration work handler checks conn state
  - Memory barriers for pre_hs_initialized flag
  - Probe work uses RCU for path references
  - hp_key tfm freed before overwrite
  - tasklet_scheduled uses atomic bit operations
  - Hysteresis state uses WRITE_ONCE/READ_ONCE
  - AEAD confidentiality limit counter enforcement

Low fixes (24):
  - HKDF label zeroization, stack buffer cleanup
  - Benchmark percentile index cast to u64
  - PMTUD cleanup ordering documented
  - RCU lock contract documented for lookup_by_psk
  - CID retire counter rolled back on alloc failure
  - Various defense-in-depth improvements

47 files changed, 2047 insertions(+), 753 deletions(-)

+++ net/tquic/bench/benchmark.c
@@ -285,7 +285,7 @@ static void compute_stats(struct tquic_bench_sample *samples, u32 count,
 
 	/* Compute percentiles */
 	for (i = 0; i < BENCH_PERCENTILES; i++) {
-		u32 idx = (percentile_indices[i] * count) / 1000;
+		u32 idx = (u32)((u64)percentile_indices[i] * count / 1000);
 		if (idx >= count)
 			idx = count - 1;
 		stats->percentiles[i] = values[idx];
